Added --masters flag to scripts-ts

+++ tools/gulptasks/scripts-ts.js
@@ -11,6 +11,8 @@
  * */
 
 const gulp = require('gulp');
+const yargs = require('yargs');
+const process = require('node:process');
 /* *
  *
  *  Tasks
@@ -26,24 +28,70 @@ const gulp = require('gulp');
 async function scriptsTS() {
     const fsLib = require('./lib/fs'),
         processLib = require('./lib/process');
+    const logLib = require('./lib/log');
+
+    const args = yargs(process.argv).argv;
 
     try {
         processLib.isRunning('scripts-ts', true);
+        if (args.masters) {
+            const { writeFile, readFile, rm } = require('node:fs/promises');
+            const tmpFileName = 'ts/tmp.tsconfig.jsonc';
 
-        fsLib.deleteDirectory('js/', true);
+            try {
+                const baseTSConfig = await readFile(
+                    'ts/no-masters.tsconfig.jsonc',
+                    'utf8'
+                );
 
-        fsLib.copyAllFiles(
-            'ts', 'js', true,
-            sourcePath => sourcePath.endsWith('.d.ts')
-        );
+                const masterFiles = args.masters.split(',');
 
-        await processLib.exec('npx tsc --build ts');
+                const pad = ' '.repeat(8);
+                const tmpTSConfig = baseTSConfig.replace(
+                    '// %% master files to compile here %%',
+                    masterFiles.map(
+                        (masterFile, i) => {
+                            // Obviously this temporary file has to be properly formatted
+                            let JSONArrayString = `"${masterFile}"`;
 
-        // Remove Dashboards
-        fsLib.deleteDirectory('js/Dashboards/', true);
-        fsLib.deleteDirectory('js/DataGrid/', true);
+                            if (i > 0) {
+                                JSONArrayString = pad + JSONArrayString;
+                            }
 
-        processLib.isRunning('scripts-ts', false);
+                            if (i < masterFiles.length - 1) {
+                                JSONArrayString += ',';
+                            }
+                            return JSONArrayString;
+                        }
+                    ).join('\n')
+                );
+
+                await writeFile(tmpFileName, tmpTSConfig, { force: true });
+                await processLib.exec(`npx tsc -P ${tmpFileName}`);
+            } finally {
+                if (!args.debug) {
+                    await rm(tmpFileName).catch(() => {
+                        logLib.warn(
+                            'Failed to remove temporary tsconfig, likely something else went wrong'
+                        );
+                    });
+                }
+            }
+        } else {
+            fsLib.deleteDirectory('js/', true);
+
+            fsLib.copyAllFiles(
+                'ts', 'js', true,
+                sourcePath => sourcePath.endsWith('.d.ts')
+            );
+
+            await processLib.exec('npx tsc --build ts');
+
+            // Remove Dashboards
+            fsLib.deleteDirectory('js/Dashboards/', true);
+            fsLib.deleteDirectory('js/DataGrid/', true);
+
+        }
     } finally {
         processLib.isRunning('scripts-ts', false);
     }
